Use weak references since we don't know lifetime
Some /sharp/ design changes here. We no longer hold the input structures
directly. Since we don't know if the underlying handle is still valid,
we can't. Instead, we hold weak references. If something attempts to use
the structure, it will build the structure through a weak reference.
This can fail if the object has already been destroyed which is what we
want.

Also, note to self, always obtain a reference when obtaining the
underlying object from a V8 Object. If a function modifies the handle,
which is supposed to be abstract to us, it won't be shown in a object
passed by value.