Add unit tests and fixes associated with them
Unit testing is super useful in our case since it's actually rather easy
to break things into pieces just to make sure we're not at least
crashing. Most things we can also behavior check but I'm not too worried
about the things we can't.

Frankly, libobs itself has no unit tests. This is a great oppurtunity to
make sure that the library itself actually works as intended as well.
For instance, I learned that scene item deletion is actually pretty slow
when there's a few hundred scene items. Not that big of a deal but might
be worth looking into eventually.