Revamp build system for ease of use
We now lazy load obs.dll. This allows us to explicitly load
obs from a designated absolute path instead of searching for the
dll in a particular directory.

We now have an appveyor build. Currently, it only supports
one artifact which is the prebuild of node-obs for the given
version of the electron runtime.

We now install into a prebuild module. Use cpack or
CMAKE_INSTALL_PREFIX to obtain the module.

We now don't compile when installing the modules and the ability
to build during install time has been removed (possibly for good).

I've removed almost all runtime dependencies. We actually have
no runtime dependencies as far as node_modules is concerned.