Apply delay load flags to all build types
Also removed an unknown and unnecessary line of code which included
a the interface target libobs a second time (and potentially
incorrectly). When interface targets are linked against, the includes
associated with that interface target are automatically added to the
include path of the building target.

In addition, appveyor wasn't building the streamlabs artifact by default