utility: Add better tested and improved Unique Id Generator
A 'Unique Id Generator' is used to uniquely identify any object quickly through the use of some form of algorithm. This implementation is a balanced mix of memory usage and speed and uses a set of ranges to determine if an id is used or is not used. Based on this set of ranges, it is incredibly quick to allocate and free ids.

The worst case complexity is O(n) with the memory usage being 40*n byte. The best case complexity is O(1) with the memory usage being 40 byte. In generic uses you can expect a mix of both and the complexity should be around O(log(n)) with the memory usage being around 4-8kb.