util-memory: Use POSIX and MSVC aligned allocators
This should fix crashes related to writing out of bounds, since the code previously could have allocated too little memory once the pointer was actually aligned. This was a simple math mistake, but using the POSIX and MSVC versions of these functions instead of writing our own should solve our issues with this.

For example, if we were to allocate 1022 bytes aligned to 16 byte and malloc returned 18472732 as the pointer, the resulting aligned pointer would be 18472752, which results in a writable area of 1020 bytes - 2 bytes less than what we wanted to allocate.