Remove shutdown synchronization
I intended for a cascading cleanup on shutdown but this turned out to be
the incorrect way to go. Instead, we should do explicit checks in the
frontend. However, we should do some explicit synchronizaton in the
bindings where crashing would occur otherwise.

I think the volmeter callbacks can be improved on here.