server/utility: Improved unique_id and unique_object_manager classes
utility::unique_id now has type definitions for the internal id for external code to use. This should help with code clarity.

utility::unique_object_manager is a new template class that can be used for singleton object management. It uses utility::unique_id, so any id that is freed is also reused later on.