server/volmeter: Fix commit a9d1c4e2c1c7764872ed5baab625f619b2d11011
A different implementation of the same idea, without the latency and instability of std::queue. Due to the way the OBS callback works (called once every audio update) and the way the Query call works, the queue method ended up with lots of audio buffered, and thus latency of up to 100ms (which is absolutely awful).

This method just uses a swap operation, setting the current data buffer to nullptr if there has been no update, which also fixes the initial issue.