server/nodeobs_display: Fix race condition in Display code
This could cause the Display code to not execute further until there was some sort of movement over the display. Things such as SetSize, SetPosition, Create and Destroy rely on this to work right.

Lesson learned in obs-vst: Don't ever use WaitMessage() if you can't guarantee that messages are sent after WaitMessage().