server/source: Slight refactor and fix tracking of private sources
An oversight caused private sources to never be tracked, due to the fact that libOBS does not run any global signals for private sources. This is likely to prevent other code or plugins to incorrectly keep a reference to something that should possibly not be visible to those. This is the fix for that by explicitly listening to the per-source destroy event.