client/utility-v8: Implement simple uv async callback runner
Our previous implementation seems to have various issues that I could not figure out, so I went with an implementation that is massively simpler to use and supports the same things. This implementation is also self-cleaning, which means that if you call finalize() it will stop the event queue, clean up and then delete itself, leaving nothing behind.

Ideally you would use this with smart pointers as data, as the code does not offer a way to set a destruction callback for T.