client/utility-v8: Fix recursive mutex lock in managed_callback
A bug in the code disallowed calling queue() on the managed_callback if the call came from within the callback. The new code instead copies or moves data over to a local variable and then works with that one instead, removing the need to keep the mutex locked for the duration of the callback execution.

This means that with this fix, you can now safely queue additional callbacks into the managed_callback and it will execute them the next time that v8 gives us control of the v8 engine. It also reduces the time code stays blocked, as the callback no longer keeps a lock on the mutex.