client: Attempt to fetch long path from short path
I don't really like this as it may still be incorrect. I actually opted into fetching from the environment variables directly and then realized that they are defined with short paths for some reason. So if the short path resolves to something other than the intended long path, then fuck you I guess. If there's a cleaner way of doing this, it isn't very clear.