cmake: Add support for running cppcheck through cmake
This allows running cppcheck directly on the project with all known defines and configurations, while offering all configuration options that are necessary. The /node_modules/ directory is automatically excluded so that we do not get errors thrown at us that aren't from our code.

This is not guaranteed to work on other platforms than windows.