Revert of PR #500
Setting m_vertexbuffer to nullptr results in crash when rebuilding device, cause vertexbufferdata is deleted, but vertexbuffer is still alive. Thus, when rebuilding device the obs tries to access the vertexbufferdata of the vertexbuffer, which results in a crash.

The alternative way to fix of original problem is presented in https://github.com/stream-labs/obs-studio/pull/186